修复droplet写数据失败
**复现步骤**

**原因和解决方案**

**影响范围**

**验证方案**

**涉及分支**

* master

**检查项**

- [ ] 需要更新依赖
- [ ] 日志打印不会造成磁盘占用空间过多
- [ ] 是共性问题(代码中存在类似问题)
- [ ] 打包成功(如果有修改Makefile)

/assign @zhuofeng

+++ ckissu/ckissu.go
@@ -237,7 +237,7 @@ var ColumnAdd610 = []*ColumnAdds{
 		Dbs:         []string{"vtap_flow_edge_port", "vtap_app_edge_port"},
 		Tables:      []string{"1m", "1m_local", "1s", "1s_local"},
 		ColumnNames: []string{"resource_gl0_type_0", "resource_gl1_type_0", "resource_gl2_type_0", "resource_gl0_type_1", "resource_gl1_type_1", "resource_gl2_type_1"},
-		ColumnType:  ckdb.UInt32,
+		ColumnType:  ckdb.UInt8,
 	},
 }
